deployer-service/requirements.txt
flask
Flask-Cors
python-dotenv
psutil
requests